requirements.txt
streamlit
python-dotenv 
litellm
pytest
openai
google-generativeai 
Pillow
tiktoken